Large Commit, read full log
Removed all old Application code. Completely redesigning it from the ground up.
Part of the reason this was done is because MARS is no longer going to be a executable application.
MARS will acts as a library that game projects will make calls to. This will allow game projects to use as much or as little of MARS as they want
without being forced to load everything.

With this change comes new problems, since MARS can no longer be "launched" there is a new project that has been added to the MARS solution.
The MARSEditor is what devs will launch if they want access to things like a level editor and asset management. This is of course optional,
devs can write their own editor using the MARS Rendering API if they want to. This is all apart of the "hands off" MARS philosophy.
The goal of MARS is to be able to use it to completely build a game off of from start to finish without the need of ever looking at the MARS source code.
At the same time, you can completely write your own "engine level" systems in your game code and without having to edit a single line of MARS source.